Extracted all the styles to the css folder
